Polish the patch.
